LPS-83052 Fix tests
