+++ modules/apps/frontend-js/frontend-js-loader-modules-extender-api/src/main/java/com/liferay/frontend/js/loader/modules/extender/npm/ModuleNameUtil.java
@@ -211,7 +211,7 @@ public class ModuleNameUtil {
 	public static String toModuleName(String fileName) {
 		String extension = FileUtil.getExtension(fileName);
 
-		if (!extension.equals(".js")) {
+		if (!extension.equals("js")) {
 			return fileName;
 		}
